feat: 폴더 구조 변화 및 input에 placeholder추가
