+++ pages/Folder/main/ActionButton/ActionButton.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from "react";
-import InputModal from "@/src/components/common/Modal/InputModal";
+import InputModal from "@/src/components/Modal/InputModal";
 import { ADD_FOLDER_ICON, ACTION_BUTTON_ICON } from "@/src/constants/image";
 import * as S from "./ActionButtonStyle";
