+++ pages/Folder/main/Popover/SelectMenu.tsx
@@ -1,6 +1,6 @@
 import { useState, forwardRef } from "react";
-import DeleteModal from "@/src/components/common/Modal/DeleteModal";
-import ListModal from "@/src/components/common/Modal/ListModal";
+import DeleteModal from "@/src/components/Modal/DeleteModal";
+import ListModal from "@/src/components/Modal/ListModal";
 import { FolderItemInfo } from "@/src/constants/types";
 import * as S from "./SelectMenuStyle";
