+++ pages/Folder/main/Toolbar/Edit/EditToolbar.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
-import InputModal from "@/src/components/common/Modal/InputModal";
-import ShareModal from "@/src/components/common/Modal/ShareModal";
-import DeleteModal from "@/src/components/common/Modal/DeleteModal";
+import InputModal from "@/src/components/Modal/InputModal";
+import ShareModal from "@/src/components/Modal/ShareModal";
+import DeleteModal from "@/src/components/Modal/DeleteModal";
 import { EDIT_SHARE, EDIT_CHANGE, EDIT_DELETE } from "@/src/constants/image";
 import * as S from "./EditToolbarStyle";
