+++ pages/_components/Footer/FooterStyle.tsx
