+++ pages/index.tsx
@@ -1,14 +1,11 @@
+import RootLayout from "./_layout";
 import Landing from "./landing/Landing";
-import Navigation from "@/src/components/Navigation/Navigation";
-import Footer from "@/src/components/Footer/Footer";
 
 const Home = () => {
   return (
-    <>
-      <Navigation />
+    <RootLayout>
       <Landing />
-      <Footer />
-    </>
+    </RootLayout>
   );
 };
